// src/shared/enums/sharepointFields.ts
export enum SharepointFields {
  company = "Company", // MMD, single value
  country = "Country",
  dateOfBirth = "Date_x0020_of_x0020_Hire",
  dateOfHire = "Date_x0020_of_x0020_Birth",
  displayName = "Display_x0020_Name",
  employee = "Employee",
  employeeCategory = "EmployeeCategoryText",
  employeeId = "Title",
  employeeMainStatus = "EmployeeMainStatusText",
  employeeStatus = "EStatus",
  employeeType = "Employee_x0020_Type",
  excludeFromOrgChart = "ExcludeFromOrgChart",
  firstName = "First_x0020_Name",
  jobTitle = "Job_x0020_Title",
  jobTitleLocal = "JobTitleLocal",
  itUser = "ItUser",
  homeCountry = "HomeCountry",
  lastName = "Last_x0020_Name",
  location = "Location",
  locationPostalCode = "Zip_x0020_Code",
  manager = "Reports_x0020_to",
  mobilePhone = "Mobile_x0020_Phone",
  orgAffiliation = "OrgAffiliation", // MMD, single value
  orgUnit = "OrgUnitText",
  orgDepartment = "DepartmentText",
  orgTeam = "TeamText",
  postOffice = "PostOffice",
  postOfficeBox = "Post_x0020_Office_x0020_Box",
  sharepointId = "Sharepoint_x0020_ID",
  streetAddress = "StreetAddress",
  workEmail = "Work_x0020_Email",
  workPhone = "Work_x0020_Phone",
}